--- ipint-payment-gateway-plugin/includes/class-wc-gateway-ipint-payment-urls.php
@@ -1,24 +1,0 @@
-<?php
-/**
- * WC_Gateway_Ipint_Payment_Url class
- *
- * @author   The Tech Makers <info@thetechmakers.com>
- * @package  WooCommerce Ipint Payments Gateway
- * @since    1.0.0
- */
-
-// Exit if accessed directly.
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
-}
-
-/**
- * Ipint Gateway.
- *
- * @class    WC_Gateway_Ipint_Settings
- * @version  1.0.1
- */
-class WC_Gateway_Ipint_Payment_Urls {
-    //add_ac
-
-}
