+++ payment.txt
@@ -1,0 +1,86 @@
+<?php
+/*
+
+Plugin Name:  WP Custom Plugin
+Plugin URI:   https://www.thetechmakers.com 
+Description:  This pulgin is developed to make payment on your store. 
+Version:      1.0.1
+Author:       The Tech Makers 
+Author URI:   https://www.thetechmakers.com
+
+*/
+
+/*
+ * This action hook registers our PHP class as a WooCommerce payment gateway
+*/
+
+/*define( API_KEY, "Ycu4foQaPiktjf1B2jb6K9SCesV9XJkC8LyhMNAkcY7e4oXv5rxPsfnAVSgNWURu" );
+define( API_SECRET, "JbjM7vtTqVZhkjkbcPBzQ9CLRVzM6dfu3CgEN2D9M5Ntu3cqrbZy9Tg95y7PYBwh" );
+define( MERCHANT_WEBSITE, "https://www.thetechmakers.com");
+define( INVOICE_CALLBACK_URL, "http://localhost/wp_plugin/payment-callback/");*/
+
+
+function initiate_payment_request() {
+    /**
+     * frontend Payment requests.
+    */
+    
+    if( isset( $_POST["pay_now"] ) ){
+
+        /**
+         * [Getting form data and posting to API request.]
+         * @var [form data]
+         */
+        $email = $_POST['email'];
+        $amount = $_POST['amount'];
+
+        $merchant_website = "https://www.thetechmakers.com";
+        $invoice_callback_url = "http://localhost/wp_plugin/payment-callback/";
+
+        if(!empty($email) && !empty($amount)){
+
+            $post_data = array('client_email_id' => $email,
+                'client_preferred_fiat_currency' => 'INR',
+                'amount' => $amount,
+                'merchant_website' => $merchant_website,
+                'invoice_callback_url' => $invoice_callback_url
+            );
+
+            $post_data = json_encode($post_data);
+            
+            $curl = curl_init();
+
+            curl_setopt_array($curl, array(
+              CURLOPT_URL => 'https://api.ipint.io:8002/checkout',
+              CURLOPT_RETURNTRANSFER => true,
+              CURLOPT_ENCODING => '',
+              CURLOPT_MAXREDIRS => 10,
+              CURLOPT_TIMEOUT => 0,
+              CURLOPT_FOLLOWLOCATION => true,
+              CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,
+              CURLOPT_CUSTOMREQUEST => 'POST',
+              CURLOPT_POSTFIELDS => $post_data,
+              CURLOPT_HTTPHEADER => array(
+                'apikey: Ycu4foQaPiktjf1B2jb6K9SCesV9XJkC8LyhMNAkcY7e4oXv5rxPsfnAVSgNWURu',
+                'Content-Type: application/json'
+              ),
+            ));
+
+            $response = curl_exec($curl);
+            
+            $response = (array) json_decode($response);
+            if(isset($response['session_id'] )){
+                $session_id = $response['session_id'];
+
+                $callback_url = "http://localhost/wp_plugin/payment-callback/?id=".$session_id; // Redirect URL.
+
+                wp_redirect( $callback_url );
+            }else{
+                echo "Something went wrong";
+            } die;
+        }
+    }
+    
+}
+
+add_action( 'wp', 'initiate_payment_request' );
