+++ readme.txt
@@ -9,7 +9,7 @@ Author URI: https://ipint.io
 Author: iPint
 Requires PHP: 7.0
 Requires at least: 4.2
-Tested up to: 4.9
+Tested up to: 6.0
 Stable tag: 1.0
 Version: 1.0 
 License: GPLv2 or later
@@ -20,6 +20,14 @@ License URI: https://www.gnu.org/licenses/gpl-2.0.html
 
 The iPint Crypto Payment Gateway plugin extends WooCommerce allowing you to take payments in crypto directly on your store or website via iPint's secure APIs with zero coding required to integrate. iPint Payment Gateway is a simple way to accept crypto payments online. With iPint you can accept Bitcoin, ETH, BCH, USDT, DOGE, LTC and many more cryptocurrencies on your platform. iPint is an end to end crypto payment processor handles crypto price fluctuations and freezes price for you so that you receive exact amount without any loss due to volatility as crypto prices changes within minutes. 
 
+The plugin is integrated with the iPint API, allowing users to pay with cryptocurrencies and preferred coins (such as BTC and ETH).
+First, you need to sign up for your merchant account with iPint. Once the sign-up process is complete, iPint will complete proper validation of the details. You can then log in and start using your account for various transactions.
+
+Merchant API credentials can be created from your iPint account. This account also allows you to add payment details based on the options you want to receive payments.
+
+Check the link for more details: https://ipint.io:8010/docs
+
+
 == Frequently Asked Questions ==
 
 = Do you have any questions? =
@@ -28,8 +36,8 @@ Drop your query to help@ipint.io , we will get back to you in 24 working hours.
 
 == Screenshots ==
 
-1. This screen shot description corresponds to Screenshot_1.png
-2. This screen shot description corresponds to Screenshot_2.png
+1. This screen shot description corresponds to assets/Screenshot_1.png
+2. This screen shot description corresponds to assets/Screenshot_2.png
 
 == Changelog ==
